market-intelligence-agent-enhanced/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv
PyJWT
passlib[bcrypt]==1.7.4
requests

# Langchain related
langchain>=0.1.9
langchain-community>=0.0.21
langchain-google-genai
langgraph

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers

# API Clients / SDKs
supabase
tavily-python
newsapi-python
alpha-vantage

# File Processing
python-multipart==0.0.9
pandas
openpyxl
unstructured[docx,excel]
Pillow
